Combine utils in form builder
