--- front/app/components/FormBuilder/utils.ts
@@ -1,65 +1,0 @@
-// services
-import {
-  ICustomFieldInputType,
-  IFlatCustomField,
-} from 'api/custom_fields/types';
-import { IPhaseData } from 'api/phases/types';
-
-// utils
-import { isNilOrError } from 'utils/helperUtils';
-
-// types
-import { MessageDescriptor } from 'utils/cl-intl';
-
-export type FormBuilderConfig = {
-  formBuilderTitle: MessageDescriptor;
-  viewFormLinkCopy: MessageDescriptor;
-  formSavedSuccessMessage: MessageDescriptor;
-  toolboxTitle?: MessageDescriptor;
-  supportArticleLink?: MessageDescriptor;
-  formEndPageLogicOption?: MessageDescriptor;
-  questionLogicHelperText?: MessageDescriptor;
-  pagesLogicHelperText?: MessageDescriptor;
-
-  toolboxFieldsToExclude: ICustomFieldInputType[];
-  formCustomFields: IFlatCustomField[] | undefined | Error;
-
-  displayBuiltInFields: boolean;
-  showStatusBadge: boolean;
-  isLogicEnabled: boolean;
-  alwaysShowCustomFields: boolean;
-  isFormPhaseSpecific: boolean;
-
-  viewFormLink?: string;
-
-  getDeletionNotice?: (projectId: string) => void;
-  getWarningNotice?: () => void;
-
-  goBackUrl?: string;
-  groupingType: 'page' | 'section';
-
-  onDownloadPDF?: () => void;
-};
-
-export const getIsPostingEnabled = (
-  phase?: IPhaseData | Error | null | undefined
-) => {
-  if (!isNilOrError(phase)) {
-    return phase.attributes.posting_enabled;
-  }
-
-  return false;
-};
-
-export const builtInFieldKeys = [
-  'title_multiloc',
-  'body_multiloc',
-  'proposed_budget',
-  'topic_ids',
-  'location_description',
-  'idea_images_attributes',
-  'idea_files_attributes',
-  'topic_ids',
-];
-
-export type BuiltInKeyType = (typeof builtInFieldKeys)[number];
